[INFRA-1053] Generate Declarative directive docs.
This is the first work in that direction - it may make sense to keep
everything in here after all rather than splitting things up as in

This generates (at least in theory) Asciidoc for the Declarative
directives that contain Describables - i.e., `agent`, `when`,
`options`, `triggers`, and `parameters`. More work is needed to
special-case `post` due to its conditions not being
Describables. Grr. I have regrets there.