Deduplicate scmGit same as GitSCM was deduplicated
Git plugin 5.0.0 added symbols in order to simplify Pipeline editing
and reading.  Instead of `$class: 'GitSCM'`, the symbol `scmGit` is used.
Instead of `$class: 'SubmoduleOption'`, the symbol `submodule` is used.
Many other examples exist.

The symbol improvement is nice for users, but it needs to be matched
with a change in the Pipeline steps doc generator.  Without this change,
the `scmGit` documentation is duplicated in many different locations
in the Pipeline steps reference.  As an example, the `multiscm` step
https://www.jenkins.io/doc/pipeline/steps/params/multiscm/ previously had
a link to a single page for `GitSCM` documentation.  Now, it embeds the
`scmGit` documentation inside the page (multiple times), duplicating
the same content that exists in several other locations.

https://issues.jenkins.io/browse/WEBSITE-809 lists pages affected by
the duplication of content that resulted from the git plugin addition
of the symbol.  Some examples from its list:

* Pipeline: Multibranch - https://www.jenkins.io/doc/pipeline/steps/workflow-multibranch/
* Pipeline: Groovy Libraries - https://www.jenkins.io/doc/pipeline/steps/pipeline-groovy-lib/

One of the key improvements for page load time and user experience with
the Pipeline steps documentation improvement project was the deduplication
of content.  We need this change to bring back that deduplication for
the git plugin documentation.