Use CODEOWNERS instead of dependabot reviewers (#562)
* Use CODEOWNERS instead of dependendabot reviewers

Dependabot is now posting this comment as a warning:

The reviewers field in the dependabot.yml file will be removed
soon. Please use the code owners file to specify reviewers for Dependabot
PRs. For more information, see this blog post:

https://github.blog/changelog/2025-04-29-dependabot-reviewers-configuration-option-being-replaced-by-code-owners/

* Use correct team as code owners

Could use jenkins-infra/docs if that is better for others